auths_core::crypto

Module signer

Signing key types.

Structs

SeedSignerKey

SignerKey implementation backed by a SecureSeed, routing through CryptoProvider.

Traits

SignerKey

A trait implemented by key types that can sign messages and return their public key.

Functions

decrypt_keypair

Decrypts a previously encrypted keypair using the configured encryption algorithm and passphrase.

encrypt_keypair

Encrypts a raw serialized keypair using the configured encryption algorithm and passphrase.

extract_seed_from_key_bytes

Extract a SecureSeed from key bytes in various formats.

load_seed_and_pubkey

Extract a SecureSeed and public key from key bytes.